+++ vle.h
@@ -87,7 +87,7 @@ typedef struct {
 
 int vle_init(vle_handle* handle, const uint8_t* buffer, const uint32_t size);
 int vle_next(vle_handle* handle, vle_t* out);
-void vle_snprint(char* str, int size, uint64_t addr, vle_t* instr);
+void vle_snprint(char* str, int size, uint32_t addr, vle_t* instr);
 
 #ifdef __cplusplus
 }
